Extract algolia step to workflow step
