+++ site/gatsby-site/package.json
@@ -154,7 +154,9 @@
     "translate-reports": "node src/scripts/translateReports.js",
     "magic-link": "npx tsx --env-file=.env src/scripts/magic-link.ts $1 $2",
     "restore-mongodb": "npx tsx --env-file=.env src/scripts/restore-mongodb.ts",
-    "import-classifications-csv": "npx tsx --env-file=.env src/scripts/import-classifications-csv.ts"
+    "import-classifications-csv": "npx tsx --env-file=.env src/scripts/import-classifications-csv.ts",
+    "algolia-update": "npx tsx --env-file=.env src/scripts/algolia-update.ts",
+    "algolia-update:ci": "npx tsx src/scripts/algolia-update.ts"
   },
   "devDependencies": {
     "@babel/plugin-proposal-export-default-from": "^7.23.3",
